Add comment about NoResultOffset. Note on typing: we generally want to use 32-bit values for offsets (which are required by some instructions in LLVM), but some of our function signatures and data structures use 64-bit types and storage locations.

+++ lib/Optimizer/Transforms/GenKernelExecution.cpp
@@ -42,7 +42,8 @@ static constexpr const char cudaqRegisterArgsCreator[] =
 static constexpr const char cudaqRegisterKernelName[] =
     "cudaqRegisterKernelName";
 
-static constexpr std::size_t NoResultOffset =
+/// This value is used to indicate that a kernel does not return a result.
+static constexpr std::uint64_t NoResultOffset =
     std::numeric_limits<std::int32_t>::max();
 
 class GenerateKernelExecution
